Fix rofi iceberg-dark theme
Rofi version 1.7.0: Iggy 2024 was released and it included a lot of
changes to the internal structure of the configuration. The old way was
slightly incompatible, and I found most of the config wasn't used
anyways.

I made it so by default the text is legible, but it may break the whole
urgent and alternate theming