Fix rofi theme
Rofi version 1.7.1: Turtley amazing! was released and it
adds a breaking change (why is this not 2.0 per semantic version?)
that moves theme configuration from being a direct property of
configuration, it now gets included similar to an @import with
@theme "theme-name" being put after the configuration block.

1.7.2: Shellebrations! was released a few days after 1.7.1, but
the Arch package has not yet been updated. This allows theme in
configuration block again, but that should now be considered
deprecated.

https://github.com/davatorium/rofi/releases/tag/1.7.1
https://github.com/davatorium/rofi/releases/tag/1.7.2